refactor: post actions
